Move feed processor-related code to a separate module
